chore: move private rpc route
