+++ src/routes/private/rpc.ts
@@ -1,10 +1,10 @@
 import { Code, ConnectError, ConnectRouter } from '@connectrpc/connect';
-import { TypeOrmError, TypeORMQueryFailedError } from '../errors';
-import { ArticlePost, SourceRequest } from '../entity';
-import { generateShortId } from '../ids';
-import createOrGetConnection from '../db';
-import { isValidHttpUrl, standardizeURL } from '../common/links';
-import { baseRpcContext } from '../common/connectRpc';
+import { TypeOrmError, TypeORMQueryFailedError } from '../../errors';
+import { ArticlePost, SourceRequest } from '../../entity';
+import { generateShortId } from '../../ids';
+import createOrGetConnection from '../../db';
+import { isValidHttpUrl, standardizeURL } from '../../common/links';
+import { baseRpcContext } from '../../common/connectRpc';
 import {
   CreatePostRequest,
   CreatePostResponse,
@@ -12,7 +12,7 @@ import {
   PostService,
 } from '@dailydotdev/schema';
 import { DataSource, FindOptionsWhere } from 'typeorm';
-import { logger } from '../logger';
+import { logger } from '../../logger';
 
 const getDuplicatePost = async ({
   req,
